more block iteration to io module
