+++ src/dolphin/io.py
@@ -1,6 +1,7 @@
+import copy
 from os import fspath
 from pathlib import Path
-from typing import List, Optional, Union
+from typing import List, Optional, Tuple, Union
 
 import numpy as np
 from osgeo import gdal
@@ -379,3 +380,234 @@ def _erode_nodata(nd_mask, buffer_pixels=25):
         out[:, j] = o
     # convert back to binary mask, and re-invert
     return ~(out > 1e-3)
+
+
+def iter_blocks(
+    filename,
+    block_shape: Tuple[int, int],
+    band=None,
+    overlaps: Tuple[int, int] = (0, 0),
+    start_offsets: Tuple[int, int] = (0, 0),
+    return_slices: bool = False,
+    skip_empty: bool = True,
+    nodata: float = np.nan,
+    nodata_mask: Optional[np.ndarray] = None,
+):
+    """Read blocks of a raster as a generator.
+
+    Parameters
+    ----------
+    filename : str or Path
+        path to raster file
+    block_shape : tuple[int, int]
+        (height, width), size of accessing blocks (default (None, None))
+    band : int, optional
+        band to read (default None, whole stack)
+    overlaps : tuple[int, int], optional
+        (row_overlap, col_overlap), number of pixels to re-include
+        after sliding the block (default (0, 0))
+    start_offsets : tuple[int, int], optional
+        (row_start, col_start), start reading from this offset
+    return_slices : bool, optional (default False)
+        return the (row, col) slice indicating the position of the current block
+    skip_empty : bool, optional (default True)
+        Skip blocks that are entirely empty (all NaNs)
+    nodata : float, optional (default np.nan)
+        Value to use for nodata to determine if a block is empty.
+        Not used if `skip_empty` is False.
+    nodata_mask : ndarray, optional
+        A boolean mask of the same shape as the raster, where True indicates
+        nodata. Ignored if `skip_empty` is False.
+        If provided, `nodata` is ignored.
+
+    Yields
+    ------
+    ndarray:
+        Current block being loaded
+
+    tuple[slice, slice]:
+        ((row_start, row_end), (col_start, col_end)) slice indicating
+        the position of the current block.
+        (Only returned if return_slices is True)
+    """
+    ds = gdal.Open(fspath(filename))
+    if band is None:
+        # Read all bands
+        read_func = ds.ReadAsArray
+    else:
+        # Read from single band
+        read_func = ds.GetRasterBand(band).ReadAsArray
+
+    # Set up the generator of ((row_start, row_end), (col_start, col_end))
+    slice_gen = slice_iterator(
+        (ds.RasterYSize, ds.RasterXSize),
+        block_shape,
+        overlaps=overlaps,
+        start_offsets=start_offsets,
+    )
+    for rows, cols in slice_gen:
+        # Skip empty blocks before reading if we have a nodata mask
+        if skip_empty and nodata_mask is not None:
+            if nodata_mask[rows, cols].all():
+                continue
+        xoff = cols.start
+        yoff = rows.start
+        xsize = cols.stop - cols.start
+        ysize = rows.stop - rows.start
+        cur_block = read_func(
+            xoff,
+            yoff,
+            xsize,
+            ysize,
+        )
+        if skip_empty:
+            # Otherwise look at the actual block we loaded
+            if np.isnan(nodata):
+                block_nodata = np.isnan(cur_block)
+            else:
+                block_nodata = cur_block == nodata
+            if np.all(block_nodata):
+                continue
+
+        if return_slices:
+            yield cur_block, (rows, cols)
+        else:
+            yield cur_block
+    ds = None
+
+
+def slice_iterator(
+    arr_shape,
+    block_shape: Tuple[int, int],
+    overlaps: Tuple[int, int] = (0, 0),
+    start_offsets: Tuple[int, int] = (0, 0),
+):
+    """Create a generator to get indexes for accessing blocks of a raster.
+
+    Parameters
+    ----------
+    arr_shape : Tuple[int, int]
+        (num_rows, num_cols), full size of array to access
+    block_shape : Tuple[int, int]
+        (height, width), size of accessing blocks
+    overlaps : Tuple[int, int]
+        (row_overlap, col_overlap), number of pixels to re-include
+        after sliding the block (default (0, 0))
+    start_offsets : Tuple[int, int]
+        Offsets to start reading from (default (0, 0))
+
+    Yields
+    ------
+    Tuple[slice, slice]
+        Iterator of (slice(row_start, row_stop), slice(col_start, col_stop))
+
+    Examples
+    --------
+        >>> list(slice_iterator((180, 250), (100, 100)))
+        [(slice(0, 100, None), slice(0, 100, None)), (slice(0, 100, None), \
+slice(100, 200, None)), (slice(0, 100, None), slice(200, 250, None)), \
+(slice(100, 180, None), slice(0, 100, None)), (slice(100, 180, None), \
+slice(100, 200, None)), (slice(100, 180, None), slice(200, 250, None))]
+        >>> list(slice_iterator((180, 250), (100, 100), overlaps=(10, 10)))
+        [(slice(0, 100, None), slice(0, 100, None)), (slice(0, 100, None), \
+slice(90, 190, None)), (slice(0, 100, None), slice(180, 250, None)), \
+(slice(90, 180, None), slice(0, 100, None)), (slice(90, 180, None), \
+slice(90, 190, None)), (slice(90, 180, None), slice(180, 250, None))]
+    """
+    rows, cols = arr_shape
+    row_off, col_off = start_offsets
+    row_overlap, col_overlap = overlaps
+    height, width = block_shape
+
+    if height is None:
+        height = rows
+    if width is None:
+        width = cols
+
+    # Check we're not moving backwards with the overlap:
+    if row_overlap >= height:
+        raise ValueError(f"row_overlap {row_overlap} must be less than {height}")
+    if col_overlap >= width:
+        raise ValueError(f"col_overlap {col_overlap} must be less than {width}")
+    while row_off < rows:
+        while col_off < cols:
+            row_end = min(row_off + height, rows)  # Dont yield something OOB
+            col_end = min(col_off + width, cols)
+            yield (slice(row_off, row_end), slice(col_off, col_end))
+
+            col_off += width
+            if col_off < cols:  # dont bring back if already at edge
+                col_off -= col_overlap
+
+        row_off += height
+        if row_off < rows:
+            row_off -= row_overlap
+        col_off = 0
+
+
+def get_max_block_shape(
+    filename, nstack: int, max_bytes: float = 64e6
+) -> Tuple[int, int]:
+    """Find shape to load from GDAL-readable `filename` with memory size < `max_bytes`.
+
+    Attempts to get an integer number of tiles from the file to avoid partial tiles.
+
+    Parameters
+    ----------
+    filename : str
+        GDAL-readable file name containing 3D dataset.
+    nstack: int
+        Number of bands in dataset.
+    max_bytes : float, optional
+        Target size of memory (in Bytes) for each block.
+        Defaults to 64e6.
+
+    Returns
+    -------
+    tuple[int]:
+        (num_rows, num_cols) shape of blocks to load from `vrt_file`
+    """
+    blockX, blockY = get_raster_block_size(filename)
+    # If it's written by line, load at least 16 lines at a time
+    blockX = max(16, blockX)
+    blockY = max(16, blockY)
+
+    ds = gdal.Open(fspath(filename))
+    shape = (ds.RasterYSize, ds.RasterXSize)
+    # get the data type from the raster
+    dt = gdal_to_numpy_type(ds.GetRasterBand(1).DataType)
+    # get the size of the data type
+    nbytes = np.dtype(dt).itemsize
+
+    full_shape = [nstack, *shape]
+    chunk_size_3d = [nstack, blockY, blockX]
+
+    # Find size of 3D chunk to load while staying at ~`max_bytes` bytes of RAM
+    chunks_per_block = max_bytes / (np.prod(chunk_size_3d) * nbytes)
+    row_chunks, col_chunks = 1, 1
+    cur_block_shape = list(copy.copy(chunk_size_3d))
+    while chunks_per_block > 1:
+        # First keep incrementing the number of columns we grab at once time
+        if col_chunks * chunk_size_3d[2] < full_shape[2]:
+            col_chunks += 1
+            cur_block_shape[2] = min(col_chunks * chunk_size_3d[2], full_shape[2])
+        # Then increase the row size if still haven't hit `max_bytes`
+        elif row_chunks * chunk_size_3d[1] < full_shape[1]:
+            row_chunks += 1
+            cur_block_shape[1] = min(row_chunks * chunk_size_3d[1], full_shape[1])
+        else:
+            break
+        chunks_per_block = max_bytes / (np.prod(cur_block_shape) * nbytes)
+    rows, cols = cur_block_shape[1:]
+    return (rows, cols)
+
+
+def get_raster_block_size(filename):
+    """Get the raster's (blockXsize, blockYsize) on disk."""
+    ds = gdal.Open(fspath(filename))
+    block_size = ds.GetRasterBand(1).GetBlockSize()
+    for i in range(2, ds.RasterCount + 1):
+        if block_size != ds.GetRasterBand(i).GetBlockSize():
+            print(f"Warning: {filename} bands have different block shapes.")
+            break
+    return block_size
