+++ src/dolphin/template_derived.vrt
@@ -5,7 +5,7 @@
         <PixelFunctionCode><![CDATA[
             import numpy as np
             def adjust(in_ar, out_ar, xoff, yoff, xsize, ysize, raster_xsize,
-                            raster_ysize, buf_radius, gt, **kwargs):
+                       raster_ysize, buf_radius, gt, **kwargs):
 
                 out_ar[:] = np.abs(in_ar[0]) * np.exp(1j * (np.angle(in_ar[0]) + np.angle(in_ar[1])))
             ]]>
