+++ tests/test_io.py
@@ -184,3 +184,128 @@ def test_get_nodata_mask(tmpdir):
     # Should still be 0
     with pytest.raises(AssertionError):
         assert m.sum() == 0
+
+
+def test_get_raster_block_sizes(raster_100_by_200, tiled_raster_100_by_200):
+    assert io.get_raster_block_size(tiled_raster_100_by_200) == [32, 32]
+    assert io.get_raster_block_size(raster_100_by_200) == [200, 1]
+    # for io.get_max_block_shape, the rasters are 8 bytes per pixel
+    # if we have 1 GB, the whole raster should fit in memory
+    bs = io.get_max_block_shape(tiled_raster_100_by_200, 1, max_bytes=1e9)
+    assert bs == (100, 200)
+
+    # for untiled, the block size is one line
+    bs = io.get_max_block_shape(raster_100_by_200, 1, max_bytes=0)
+    # The function forces at least 16 lines to be read at a time
+    assert bs == (16, 200)
+    bs = io.get_max_block_shape(raster_100_by_200, 1, max_bytes=8 * 17 * 200)
+    assert bs == (32, 200)
+
+    # Pretend we have a stack of 10 images
+    nstack = 10
+    # one tile should be 8 * 32 * 32 * 10 = 81920 bytes
+    bytes_per_tile = 8 * 32 * 32 * nstack
+    bs = io.get_max_block_shape(
+        tiled_raster_100_by_200, nstack, max_bytes=bytes_per_tile
+    )
+    assert bs == (32, 32)
+
+    # with a little more, we should get 2 tiles
+    bs = io.get_max_block_shape(
+        tiled_raster_100_by_200, nstack, max_bytes=1 + bytes_per_tile
+    )
+    assert bs == (32, 64)
+
+    # 200 / 32 = 6.25, so with 7, it should add a new row
+    bs = io.get_max_block_shape(
+        tiled_raster_100_by_200, nstack, max_bytes=7 * bytes_per_tile
+    )
+    assert bs == (64, 200)
+
+
+def test_iter_blocks(tiled_raster_100_by_200):
+    # Try the whole raster
+    bs = io.get_max_block_shape(tiled_raster_100_by_200, 1, max_bytes=1e9)
+    blocks = list(io.iter_blocks(tiled_raster_100_by_200, bs, band=1))
+    assert len(blocks) == 1
+    assert blocks[0].shape == (100, 200)
+
+    # now one block at a time
+    max_bytes = 8 * 32 * 32
+    bs = io.get_max_block_shape(tiled_raster_100_by_200, 1, max_bytes=max_bytes)
+    blocks = list(io.iter_blocks(tiled_raster_100_by_200, bs, band=1))
+    row_blocks = 100 // 32 + 1
+    col_blocks = 200 // 32 + 1
+    expected_num_blocks = row_blocks * col_blocks
+    assert len(blocks) == expected_num_blocks
+    assert blocks[0].shape == (32, 32)
+    # at the ends, the blocks are smaller
+    assert blocks[6].shape == (32, 8)
+    assert blocks[-1].shape == (4, 8)
+
+
+def test_iter_nodata(
+    raster_with_nan,
+    raster_with_nan_block,
+    raster_with_zero_block,
+    tiled_raster_100_by_200,
+):
+    # load one block at a time
+    max_bytes = 8 * 32 * 32
+    bs = io.get_max_block_shape(tiled_raster_100_by_200, 1, max_bytes=max_bytes)
+    blocks = list(io.iter_blocks(tiled_raster_100_by_200, bs, band=1))
+    row_blocks = 100 // 32 + 1
+    col_blocks = 200 // 32 + 1
+    expected_num_blocks = row_blocks * col_blocks
+    assert len(blocks) == expected_num_blocks
+    assert blocks[0].shape == (32, 32)
+
+    # One nan should be fine, will get loaded
+    blocks = list(
+        io.iter_blocks(raster_with_nan, bs, band=1, skip_empty=True, nodata=np.nan)
+    )
+    assert len(blocks) == expected_num_blocks
+
+    # Now check entire block for a skipped block
+    blocks = list(
+        io.iter_blocks(
+            raster_with_nan_block, bs, band=1, skip_empty=True, nodata=np.nan
+        )
+    )
+    assert len(blocks) == expected_num_blocks - 1
+
+    # Now check entire block for a skipped block
+    blocks = list(
+        io.iter_blocks(raster_with_zero_block, bs, band=1, skip_empty=True, nodata=0)
+    )
+    assert len(blocks) == expected_num_blocks - 1
+
+
+def test_iter_blocks_nodata_mask(tiled_raster_100_by_200):
+    # load one block at a time
+    max_bytes = 8 * 32 * 32
+    bs = io.get_max_block_shape(tiled_raster_100_by_200, 1, max_bytes=max_bytes)
+    blocks = list(io.iter_blocks(tiled_raster_100_by_200, bs, band=1))
+    row_blocks = 100 // 32 + 1
+    col_blocks = 200 // 32 + 1
+    expected_num_blocks = row_blocks * col_blocks
+    assert len(blocks) == expected_num_blocks
+
+    nodata_mask = np.zeros((100, 200), dtype=np.bool)
+    nodata_mask[:5, :5] = True
+    # non-full-block should still all be loaded nan should be fine, will get loaded
+    blocks = list(
+        io.iter_blocks(
+            tiled_raster_100_by_200, bs, skip_empty=True, nodata_mask=nodata_mask
+        )
+    )
+    assert len(blocks) == expected_num_blocks
+
+    nodata_mask[:32, :32] = True
+    # non-full-block should still all be loaded nan should be fine, will get loaded
+    blocks = list(
+        io.iter_blocks(
+            tiled_raster_100_by_200, bs, skip_empty=True, nodata_mask=nodata_mask
+        )
+    )
+    assert len(blocks) == expected_num_blocks - 1
